+++ app/src/main/java/com/github/bumblebee202111/doubean/feature/groups/common/TopicItemWithGroupAndroidView.kt
@@ -9,6 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidViewBinding
@@ -17,6 +18,8 @@ import com.github.bumblebee202111.doubean.R
 import com.github.bumblebee202111.doubean.databinding.ListItemPostNotificationBinding
 import com.github.bumblebee202111.doubean.model.TopicItemWithGroup
 import com.github.bumblebee202111.doubean.ui.common.UserProfileImage
+import com.github.bumblebee202111.doubean.ui.component.DateTimeText
+import com.github.bumblebee202111.doubean.util.abbreviatedDateTimeString
 
 @Composable
 fun TopicItemWithGroupAndroidView(
@@ -34,6 +37,13 @@ fun TopicItemWithGroupAndroidView(
                 size = dimensionResource(id = R.dimen.icon_size_extra_small)
             )
         }
+
+        created.setContent {
+            topicItemWithGroup?.created?.let {
+                DateTimeText(text = it.abbreviatedDateTimeString(LocalContext.current))
+            }
+        }
+
         cover.setContent {
             topicItemWithGroup?.coverUrl?.let {
                 AsyncImage(
@@ -46,15 +56,23 @@ fun TopicItemWithGroupAndroidView(
             }
 
         }
+
         card.setOnClickListener {
             topicItemWithGroup?.let { navigateToTopic(it.id) }
         }
+
         commentIcon.setContent {
             Icon(
                 imageVector = Icons.AutoMirrored.Outlined.Comment,
                 contentDescription = null,
                 modifier = Modifier.size(dimensionResource(id = R.dimen.icon_size_extra_small))
             )
+        }
+
+        lastUpdated.setContent {
+            topicItemWithGroup?.lastUpdated?.let {
+                DateTimeText(text = it.abbreviatedDateTimeString(LocalContext.current))
+            }
         }
     }
 }