+++ app/src/main/java/com/github/bumblebee202111/doubean/ui/component/DateTimeText.kt
@@ -1,0 +1,15 @@
+package com.github.bumblebee202111.doubean.ui.component
+
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun DateTimeText(modifier: Modifier = Modifier, text: String) {
+    Text(
+        text = text,
+        modifier = modifier,
+        style = MaterialTheme.typography.bodyMedium
+    )
+}